+++ UltimateCastbars/Backend/DefaultDB/Target_DB.lua
@@ -485,6 +485,7 @@ local target_db = {
                 channel = 0.5,
                 empowered = 0.5,
             },
+            channelError = 100,
         },
     },